fixing integrity error with hosted jquery
The original file was hosted by a cdn and moved here, so the integrity (sha256) is
not valid. This will test remove it. It could also be the case that netlify does not
require the baseurl (if it just deploys from _site) so I will make an adjustment after thi
s if that turns out to be the case!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>