fixing bug with link parsing
The current bug (issue #3) is happening because we always replace the link text
(typically a markdown reference) with an updated parsed link. The issue is
that this reference to the markdown file can happen multiple times and earlier,
so the wrong text is replaced. This fix uses a regular expression to get the
entire a href content, and then assembles a new link and replaces the first
instance of that entire block.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>